+++ src/pattern_test.cc
@@ -71,12 +71,14 @@ test_case(Pattern, SettersAndGetters) {
  * something went wrong via the standard error output.
  */
 test_case(Pattern, Factory) {
-  const int ROWS = 64;
-  const int TRACKS = 8;
+  const unsigned int ROWS = 64;
+  const unsigned int TRACKS = 8;
 
   // Create a mock-up of the tracks list.
   MockTracks tracks;
-  for (int i = 0; i < TRACKS; i++) { tracks.push_back(new MockTrack()); }
+  for (unsigned int i = 0; i < TRACKS; i++) {
+    tracks.push_back(new MockTrack());
+  }
 
   /*
    * Normal operations for creating a new pattern based on the track list and
@@ -92,7 +94,7 @@ test_case(Pattern, Factory) {
 
   // Try again but with mismatching rows-param and pattern_rows contents.
   MockPatternRows pattern_rows;
-  for (int i = 0; i < ROWS - 1; i++) {
+  for (unsigned int i = 0; i < ROWS - 1; i++) {
     pattern_rows.push_back(new MockPatternRow());
   }
   assert_stderr_eq("ERROR: The number of pattern rows provided does not "
@@ -104,7 +106,9 @@ test_case(Pattern, Factory) {
   assert_eq(NULL, pattern);
 
   // Clean-up the mocked tracks
-  for (int i = 0; i < TRACKS; i++) { delete tracks.back(); tracks.pop_back(); }
+  for (unsigned int i = 0; i < TRACKS; i++) {
+    delete tracks.back(); tracks.pop_back();
+  }
 }
 
 
@@ -114,21 +118,21 @@ test_case(Pattern, Factory) {
  * Make sure that rows are added when the factory method AddRows is called.
  */
 test_case(Pattern, Factory_adding_rows_shall_add_rows) {
-  const int ROWS= 2;
-  const int TRACKS = 8;
+  const unsigned int ROWS= 2;
+  const unsigned int TRACKS = 8;
 
   MockTracks tracks;
   MockTrack track;
   Pattern *pattern = PatternFactoryMock::New(&tracks, ROWS);
 
-  for (int i = 0; i < TRACKS; i++) { tracks.push_back(&track); }
+  for (unsigned int i = 0; i < TRACKS; i++) { tracks.push_back(&track); }
 
   PatternFactoryMock::AddRows(&tracks, pattern, 6);
 
   MockPatternRows *rows =
     dynamic_cast<MockPatternRows*>(pattern->get_pattern_rows());
 
-  assert_eq(8, rows->size());
+  assert_eq(static_cast<unsigned int>(8), rows->size());
 }
 
 
@@ -139,19 +143,19 @@ test_case(Pattern, Factory_adding_rows_shall_add_rows) {
  * called.
  */
 test_case(Pattern, Factory_removing_rows_shall_remove_rows) {
-  const int ROWS= 8;
-  const int TRACKS = 8;
+  const unsigned int ROWS= 8;
+  const unsigned int TRACKS = 8;
 
   MockTracks tracks;
   MockTrack track;
   Pattern *pattern = PatternFactoryMock::New(&tracks, ROWS);
 
-  for (int i = 0; i < TRACKS; i++) { tracks.push_back(&track); }
+  for (unsigned int i = 0; i < TRACKS; i++) { tracks.push_back(&track); }
 
   PatternFactoryMock::RemoveRows(pattern, 2);
 
   MockPatternRows *rows =
     dynamic_cast<MockPatternRows*>(pattern->get_pattern_rows());
 
-  assert_eq(6, rows->size());
+  assert_eq(static_cast<unsigned int>(6), rows->size());
 }
